Refactoring helpers and components to own files
