+++ src/Popover/helpers/resolvePopoverAnchor.flow.js
@@ -1,0 +1,9 @@
+// @flow
+import type { Anchor, Dimensions } from "../index.flow";
+
+export type StyledAnchor = {|
+  ...Anchor,
+  ...Dimensions,
+|};
+
+declare export default StyledAnchor;
