+++ src/Popover/helpers/resolvePopoverPositin.flow.js
@@ -1,0 +1,9 @@
+// @flow
+import type { Position, Dimensions } from "../index.flow";
+
+export type StyledPosition = {|
+  ...Position,
+  ...Dimensions,
+|};
+
+declare export default StyledPosition;
